pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-matcher"
version = "0.1.0"
description = "批量快递地址与高德POI匹配的离线流水线"
authors = [{name = "Codex Assistant"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2",
    "openpyxl>=3.1",
    "pydantic>=2.8",
    "rapidfuzz>=3.9",
    "jieba>=0.42",
    "pypinyin>=0.49",
    "typer>=0.12",
    "pyyaml>=6.0.1",
    "loguru>=0.7",
    "tqdm>=4.66"
]

[project.optional-dependencies]
vector = [
    "sentence-transformers>=2.6",
    "scikit-learn>=1.5"
]

[project.scripts]
geo-matcher = "geo_matcher.cli:main"
